several suggestions for the latest dev version
