+++ README.md
@@ -1,0 +1,35 @@
+## Sensitivity of TALYS Model Parameters for Medical Isotope Production Cross Sections
+
+This repository contains Python scripts designed for a sensitivity study of the model parameters used in the nuclear reaction model code [TALYS](https://nds.iaea.org/talys/). The goal is to optimize the production cross sections of medical isotopes. These scripts provide functionality to:
+
+- Perform TALYS calculations for isotopes listed in the IAEA Medical Isotope Evaluation.
+- Read and analyze output files generated by TALYS.
+- Obtain optimized chi-square values using the least-squares method.
+
+By leveraging these tools, users can systematically refine TALYS parameters to improve the reliability of medical isotope production data.
+
+## How to run the code
+First, you need to install necessary Python packages by running
+```
+pip install -r requirements.txt
+```
+
+To execute the sensitivity study, simply run the following command in your terminal:
+```
+python calc.py
+```
+
+## Documentation
+The documentation is available online in IAEA Nuclear Data Section website as [IAEA-NDS-0249](https://doi.org/10.61092/iaea.8j2z-vz4q).
+
+
+## Citation
+If you want to cite this script, please use the following reference:
+```
+S. Nakamizo and A. Koning, "Sensitivity of TALYS Model Parameters for Medical Isotope Production Cross Sections", Technical Report IAEA-NDS-0249, International Atomic Energy Agency, Vienna, 2024.
+```
+
+## License
+This code is distributed under the MIT license augmented by an IAEA clause, see the accompanying license file for more information.
+
+Copyright (c) International Atomic Energy Agency (IAEA)
