+++ config.py
@@ -6,16 +6,16 @@ XS_THRESHOLD_PERCENTAGE = 5 # %
 FIT = "n"
 
 
-TALYS_PATH = "/home/shuri/Documents/talys"
-CALC_PATH = f"/home/shuri/Documents/talys/medical/calc026_a_fit_{FIT}"
-EXFORTABLES_PATH = "/home/shuri/Documents/exfortables"
+# HOME_DIR = "/home/shuri/Documents" 
+HOME_DIR = "/Users/okumuras/Documents/codes" 
+TALYS_PATH = os.path.join(HOME_DIR, "talys")
+CALC_PATH = os.path.join(HOME_DIR, f"medical/calc026_a_fit_{FIT}")
+EXFORTABLES_PATH =  os.path.join(HOME_DIR, "exfortables")
 TALYS_INP_FILE_NAME = "talys.inp"
-# GNUPLOT_OUTPUT_DIRECTORY_NAME = gnuplot_output_
 
 
-IAEA_MEDICAL_LIST = "/home/shuri/Documents/calculate_chi_square_total/data/IAEA_medical_isotope.dat"
-
-SCORE_JSON_PATH = "/home/shuri/Documents/talys/json"
+IAEA_MEDICAL_LIST = "./data/IAEA_medical_isotope.dat"
+SCORE_JSON_PATH = os.path.join(HOME_DIR, "talys/json")
 
 ENERGY_RANGE_MIN = 1.0
 ENERGY_RANGE_MAX = 60.0
@@ -26,5 +26,4 @@ ENERGY_STEP = 0.5
 ## Number of parallel processes
 N = 5
 TIMEOUT_SECONDS = 2000
-
 LATEX_DOCUMENT_FILENAME = f"{os.path.basename(CALC_PATH)}.tex"
