--- delete/create_projectile_files_ga.py
@@ -1,115 +1,0 @@
-def create_projectile_file1(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 1\n")
-        f.write("colenhance n\n")
-        f.write("rwdadjust p 1.02490\n")
-        f.write("awdadjust p 0.97348\n")
-        f.write("rvadjust n 1.02521\n")
-        f.write("gadjust 32 70 1.10125\n")
-        f.write("gadjust 32 69 1.20392\n")
-        f.write("gadjust 31 69 0.75348\n")
-        
-
-def create_projectile_file2(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 1\n")
-        f.write("colenhance y\n")
-        f.write("rwdadjust p 1.05947\n")
-        f.write("awdadjust p 1.01703\n")
-        f.write("rvadjust n 0.97366\n")
-        f.write("gadjust 32 70 1.20796\n")
-        f.write("gadjust 32 69 1.38259\n")
-        f.write("gadjust 31 69 0.99756\n")
-
-def create_projectile_file3(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 2\n")
-        f.write("colenhance n\n")
-        f.write("rwdadjust p 0.98112\n")
-        f.write("awdadjust p 0.96744\n")
-        f.write("rvadjust n 1.02381\n")
-        f.write("gadjust 32 70 1.09024\n")
-        f.write("gadjust 32 69 1.03924\n")
-        f.write("gadjust 31 69 1.29751\n")
-
-def create_projectile_file4(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 2\n")
-        f.write("colenhance y\n")
-        f.write("rwdadjust p 1.00893\n")
-        f.write("awdadjust p 0.97503\n")
-        f.write("rvadjust n 1.00982\n")
-        f.write("gadjust 32 70 1.06840\n")
-        f.write("gadjust 32 69 0.85346\n")
-        f.write("gadjust 31 69 1.18839\n")
-
-def create_projectile_file5(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 5\n")
-        f.write("colenhance n\n")
-        f.write("rwdadjust p 0.97331\n")
-        f.write("awdadjust p 1.01511\n")
-        f.write("rvadjust n 1.03244\n")
-        f.write("gadjust 32 70 1.00896\n")
-        f.write("gadjust 32 69 0.92886\n")
-        f.write("gadjust 31 69 1.10381\n")
