--- delete/create_projectile_files_o.py
@@ -1,115 +1,0 @@
-def create_projectile_file1(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 1\n")
-        f.write("colenhance n\n")
-        f.write("rwdadjust p 0.94295\n")
-        f.write("awdadjust p 0.76930\n")
-        f.write("rvadjust n 0.97610\n")
-        f.write("gadjust 9 19 0.85961\n")
-        f.write("gadjust 9 18 1.01027\n")
-        f.write("gadjust 8 18 1.02088\n")
-        
-
-def create_projectile_file2(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 1\n")
-        f.write("colenhance y\n")
-        f.write("rwdadjust p 0.91909\n")
-        f.write("awdadjust p 0.68433\n")
-        f.write("rvadjust n 1.03010\n")
-        f.write("gadjust 9 19 1.01797\n")
-        f.write("gadjust 9 18 1.03134\n")
-        f.write("gadjust 8 18 0.95660\n")
-
-def create_projectile_file3(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 2\n")
-        f.write("colenhance n\n")
-        f.write("rwdadjust p 0.87793\n")
-        f.write("awdadjust p 0.83804\n")
-        f.write("rvadjust n 0.99799\n")
-        f.write("gadjust 9 19 1.08395\n")
-        f.write("gadjust 9 18 1.09820\n")
-        f.write("gadjust 8 18 0.97648\n")
-
-def create_projectile_file4(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 2\n")
-        f.write("colenhance y\n")
-        f.write("rwdadjust p 0.87701\n")
-        f.write("awdadjust p 0.88110\n")
-        f.write("rvadjust n 0.98426\n")
-        f.write("gadjust 9 19 1.18091\n")
-        f.write("gadjust 9 18 1.06672\n")
-        f.write("gadjust 8 18 1.25692\n")
-
-def create_projectile_file5(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 5\n")
-        f.write("colenhance n\n")
-        f.write("rwdadjust p 0.77953\n")
-        f.write("awdadjust p 1.01123\n")
-        f.write("rvadjust n 1.00595\n")
-        f.write("gadjust 9 19 1.24828\n")
-        f.write("gadjust 9 18 1.37736\n")
-        f.write("gadjust 8 18 1.12684\n")
