--- delete/create_projectile_files_te.py
@@ -1,115 +1,0 @@
-def create_projectile_file1(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 1\n")
-        f.write("colenhance n\n")
-        f.write("rwdadjust p 0.96744\n")
-        f.write("awdadjust p 0.93559\n")
-        f.write("rvadjust n 0.99936\n")
-        f.write("gadjust 53 125 1.01299\n")
-        f.write("gadjust 53 124 0.97121\n")
-        f.write("gadjust 52 124 1.12173\n")
-        
-
-def create_projectile_file2(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 1\n")
-        f.write("colenhance y\n")
-        f.write("rwdadjust p 0.98112\n")
-        f.write("awdadjust p 0.92020\n")
-        f.write("rvadjust n 1.01959\n")
-        f.write("gadjust 53 125 1.01616\n")
-        f.write("gadjust 53 124 0.83874\n")
-        f.write("gadjust 52 124 1.40329\n")
-
-def create_projectile_file3(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 2\n")
-        f.write("colenhance n\n")
-        f.write("rwdadjust p 0.92675\n")
-        f.write("awdadjust p 0.97057\n")
-        f.write("rvadjust n 1.00869\n")
-        f.write("gadjust 53 125 0.92111\n")
-        f.write("gadjust 53 124 0.87540\n")
-        f.write("gadjust 52 124 1.24187\n")
-
-def create_projectile_file4(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 2\n")
-        f.write("colenhance y\n")
-        f.write("rwdadjust p 0.98112\n")
-        f.write("awdadjust p 0.86932\n")
-        f.write("rvadjust n 1.03982\n")
-        f.write("gadjust 53 125 0.96149\n")
-        f.write("gadjust 53 124 0.87139\n")
-        f.write("gadjust 52 124 1.49098\n")
-
-def create_projectile_file5(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 5\n")
-        f.write("colenhance n\n")
-        f.write("rwdadjust p 0.98112\n")
-        f.write("awdadjust p 0.84242\n")
-        f.write("rvadjust n 1.05059\n")
-        f.write("gadjust 53 125 0.89791\n")
-        f.write("gadjust 53 124 0.70930\n")
-        f.write("gadjust 52 124 1.25060\n")
