--- delete/create_projectile_files_y.py
@@ -1,115 +1,0 @@
-def create_projectile_file1(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 1\n")
-        f.write("colenhance n\n")
-        f.write("rwdadjust p 1.01244\n")
-        f.write("awdadjust p 0.92997\n")
-        f.write("rvadjust n 0.92437\n")
-        f.write("gadjust 40 90 1.08918\n")
-        f.write("gadjust 40 89 1.16020\n")
-        f.write("gadjust 39 89 0.99689\n")
-        
-
-def create_projectile_file2(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 1\n")
-        f.write("colenhance y\n")
-        f.write("rwdadjust p 0.95307\n")
-        f.write("awdadjust p 0.69077\n")
-        f.write("rvadjust n 0.70940\n")
-        f.write("gadjust 40 90 1.32264\n")
-        f.write("gadjust 40 89 1.94442\n")
-        f.write("gadjust 39 89 0.20806\n")
-
-def create_projectile_file3(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 2\n")
-        f.write("colenhance n\n")
-        f.write("rwdadjust p 0.93951\n")
-        f.write("awdadjust p 0.89624\n")
-        f.write("rvadjust n 0.93520\n")
-        f.write("gadjust 40 90 1.12391\n")
-        f.write("gadjust 40 89 1.22095\n")
-        f.write("gadjust 39 89 0.91306\n")
-
-def create_projectile_file4(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 2\n")
-        f.write("colenhance y\n")
-        f.write("rwdadjust p 0.92715\n")
-        f.write("awdadjust p 0.99703\n")
-        f.write("rvadjust n 0.92399\n")
-        f.write("gadjust 40 90 1.10283\n")
-        f.write("gadjust 40 89 1.07922\n")
-        f.write("gadjust 39 89 1.10589\n")
-
-def create_projectile_file5(projectile, element, mass, energy, output_file):
-    with open(output_file, 'w') as f:
-        f.write("#\n")
-        f.write(f"#  {projectile}-{element}\n")
-        f.write("#\n")
-        f.write("# General\n")
-        f.write("#\n")
-        f.write(f"projectile {projectile}\n")
-        f.write(f"element {element}\n")
-        f.write(f"mass {mass}\n")
-        f.write(f"energy {energy}\n")
-        f.write("#\n")
-        f.write("# Parameters\n")
-        f.write("#\n")
-        f.write("ldmodel 5\n")
-        f.write("colenhance n\n")
-        f.write("rwdadjust p 0.89916\n")
-        f.write("awdadjust p 0.96301\n")
-        f.write("rvadjust n 0.80958\n")
-        f.write("gadjust 40 90 1.29725\n")
-        f.write("gadjust 40 89 1.69950\n")
-        f.write("gadjust 39 89 0.41586\n")
