+++ requirements.txt
@@ -1,0 +1,4 @@
+appdirs==1.4.4
+lark==1.2.2
+numpy==2.2.0
+tqdm==4.67.1
